Update on Window
Implementation of SDL_Texture and SDL_Surface. Needs SDL_QueueTexture (or something) to actually display the texture.
Everything is being called from JetEngine.cpp to Window.cpp but later the best will maybe be implement the textures part of the render on something more related to that, as the Level class or something alike.