Started Refactoring and Created new classes to reorganize the whole Solution - Functioning Backup
Added classes to JetEngine (.h and .cpp)
- Animation
- Pawn

Added classes to Xenon Clone Game (.h and .cpp)
- Companion
- Drone
- Enemies
- Hazards
- Loner
- MetalAsteroid
- Projectiles
- Rusher
- Spaceship
- Stone Asteroid

All headers have the #ifndef ... include guards as of now.
Created, for each project a "Previous" filter (folder) to store the code that was already implemented, in order to adapt it to new classes.

Also, outside of the Solution, was put the Project 2 PDF and a PNG of the current Programming Plan:
- Red -> JetEngine Classes, with the white being the variables I deemed necessary
- Blue -> Xenon Game Clone superclasses, with white being the derived classes

- Yellow, the overall Game Loop for the Game.
The idea, as of now, is to each class has implemented its own update, and then the main file, either the "exe" or the Engine main file, call it. I'm still trying to decide on that. Also trying to decide if the Companion class derives from Player (which I'm renaming of "Pawn", as mentioned on the beginning, or GameObject.