Implemented companions and the rest of missiles
Its missing the Damage logic with all that is damage related, and its missing the box 2D collisions debug and detection.
apparently, Box2d has shapes that can be sensors. Sensors detect simply passing through other shapes without collisions.
A Body can have as many collisions as you want, so you can for example create a normal shape and a sensor shape for the same body if you want.
Sensors do not detect sensors, only normal shapes.
To be able to detect what type of shape the sensors are detecting, you can use layers and compare layers.
For debugging, you can use the body data to draw a square with SDL or OpenGl.

For sensors you need to set isSensor to true and enableSensorEvents to true when you create the shape.

The detection between the different actors has to be implemented by you. Each "Actor" has its layer