Box2D updated
FOR SOME REASON, THE COLLISIONS ARE NOT BEING DETECTED.
ALSO, TAGS AND TAG COMPARISONS LOGIC IS MISSING.
THE RIGIDBODIES ARE NOT BEING UPDATED

The spaceship can debug the rigidbody created, but for some reason the other objects can't output debug renders.

Its missing the PowerUps in-game, the collision logic, the movement of rigidbodies, the tags for the collisions, and the meteorites.